refactor(Follow): 팔로워 음수 방어 로직 추가
- 커스텀 예외 추가: FollowerCountCannotBeNegativeException

+++ src/main/java/com/codeit/mopl/domain/follow/service/FollowService.java
@@ -11,10 +11,7 @@ import com.codeit.mopl.domain.user.entity.User;
 import com.codeit.mopl.domain.user.repository.UserRepository;
 import com.codeit.mopl.event.event.FollowerDecreaseEvent;
 import com.codeit.mopl.event.event.FollowerIncreaseEvent;
-import com.codeit.mopl.exception.follow.FollowDeleteForbiddenException;
-import com.codeit.mopl.exception.follow.FollowDuplicateException;
-import com.codeit.mopl.exception.follow.FollowNotFoundException;
-import com.codeit.mopl.exception.follow.FollowSelfProhibitedException;
+import com.codeit.mopl.exception.follow.*;
 import com.codeit.mopl.exception.user.UserErrorCode;
 import com.codeit.mopl.exception.user.UserIdIsNullException;
 import com.codeit.mopl.exception.user.UserNotFoundException;
@@ -93,14 +90,21 @@ public class FollowService {
         log.info("[팔로우 관리] 팔로우 삭제 시작 - followId: {}", followId);
         Follow follow = followRepository.findById(followId)
                 .orElseThrow(() -> FollowNotFoundException.withId(followId));
-        UUID followerId = follow.getFollower().getId();
-        
+
         // 팔로우한 본인이 아니면 팔로우 삭제 불가능
+        UUID followerId = follow.getFollower().getId();
         if (!followerId.equals(requesterId)) {
             throw FollowDeleteForbiddenException.withIds(followId, followerId, requesterId);
         }
 
+        // followerCount는 음수가 될 수 없음
+        User followee = follow.getFollowee();
         UUID followeeId = follow.getFollowee().getId();
+        long followerCount = followee.getFollowerCount();
+        if (followerCount-1 <= 0) {
+            log.error("[팔로우 관리] 팔로우 삭제 중단 - 팔로워 수가 이미 0입니다.");
+            throw FollowerCountCannotBeNegativeException.withId(followId);
+        }
 
         followRepository.deleteById(followId);
         eventPublisher.publishEvent(new FollowerDecreaseEvent(followeeId));
